+++ BALSAMIC/constants/workflow_params.py
@@ -107,6 +107,9 @@ WORKFLOW_PARAMS = {
     "bam_post_processing": {
         "manta_max_base_quality": 70,
     },
+    "bed_pre_processing": {
+        "minimum_region_size": 100,
+    },
     "common": {
         "header_per_lane": "'@RG\\tID:{fastq_pattern}\\tSM:{sample_type}\\tPL:ILLUMINAi'",
         "header_per_sample": "'@RG\\tID:{sample}\\tSM:{sample_type}\\tPL:ILLUMINAi'",