+++ BALSAMIC/utils/cli.py
@@ -215,6 +215,44 @@ def bioinfo_tool_version_conda(
     return conda_bioinfo_version
 
 
+def get_gens_references(
+    genome_interval: Optional[str],
+    gens_coverage_pon: Optional[str],
+    gnomad_min_af5: Optional[str],
+    panel_bed: Optional[str],
+) -> Dict[str, str]:
+    """
+    Assigns reference-files required for GENS if they have been supplied, else exists with error message.
+
+    Args:
+        genome_interval: Optional[str]. Coverage-regions. (required for WGS GENS)
+        gens_coverage_pon: Optional[str]. PON for GATK CollectReadCounts. (required for WGS GENS)
+        gnomad_min_af5: Optional[str] gnomad VCF filtered to keep variants above 5% VAF (required for WGS and TGA GENS).
+        panel_bed: Optional[str] Bedfile supplied for TGA analyses.
+
+    Returns:
+         Dict[str, str] with paths to GENS reference files
+    """
+
+    if panel_bed and gnomad_min_af5:
+        return {"gnomad_min_af5": gnomad_min_af5}
+
+    if gnomad_min_af5 and genome_interval and gens_coverage_pon:
+        return {
+            "genome_interval": genome_interval,
+            "gens_coverage_pon": gens_coverage_pon,
+            "gnomad_min_af5": gnomad_min_af5,
+        }
+
+    error_message = (
+        "GENS reference file is always required to run BALSAMIC."
+        "WGS requires arguments: genome_interval, gens_coverage_pon, gnomad_min_af5"
+        "TGA requires argument: gnomad_min_af5"
+    )
+    LOG.error(error_message)
+    raise BalsamicError(error_message)
+
+
 def get_bioinfo_tools_version(
     bioinfo_tools: dict, container_conda_env_path: Path
 ) -> dict: